dodano kompajliranje preko emscriptena i prilagođen kod
git-svn-id: https://r7/svn/RenderDragon@39 4bef5c89-85b5-4243-b844-44d7e83bc23a

+++ graphics/objects/vertex_buffer.cpp
@@ -28,7 +28,7 @@ std::vector<byte> CVertexBuffer::InterleaveVertexData(const SVertexFormat& fmt, 
 	}
 	
 	std::vector<byte> databytes; databytes.resize(vertexCount*vtxbytesize);
-	for(auto& b : databytes) b = 0x00;
+	for(auto& b : databytes) b = (byte)0x00;
 
 	{
 		byte* dataptr = databytes.data();
